🎈 perf(settings): Improve Settings
I realized we can scrape the fields and default values from the HTML instead of declaring them individually.

Since I use the form, and I convert the values to an object, and I do this multiple times in jobForm and settings, I made a generic function in utils and now call it.